backend/requirements.txt
flask
flask-cors
flask-socketio
eventlet
python-dotenv
requests
openai
